[ICU-13275] Merge warning cleanup fixes into trunk X-Commit-URL: https://ssl.icu-project.org/trac/changeset/40283 Former-commit-id: 28b6d45d1369ab9888153629e3470bb36b4ae362

+++ icu4c/source/common/uhash.cpp
@@ -844,7 +844,7 @@ uhash_hashUChars(const UHashTok key) {
 U_CAPI int32_t U_EXPORT2
 uhash_hashChars(const UHashTok key) {
     const char *s = (const char *)key.pointer;
-    return s == NULL ? 0 : ustr_hashCharsN(s, uprv_strlen(s));
+    return s == NULL ? 0 : static_cast<int32_t>(ustr_hashCharsN(s, uprv_strlen(s)));
 }
 
 U_CAPI int32_t U_EXPORT2
